feat: added global function prototypes

+++ lib/2d/src/rect.c
@@ -1,8 +1,7 @@
-#include <stdio.h>
-#include "../inc/rect.h"
+#include "../../lib.h"
 #include "../../common/common.h"
 
-rect_t *rect(int x1, int y1, unsigned int width, unsigned int height)
+rect_t *rect(int x1, int y1, size_t width, size_t height)
 {
     int __x1 = x1, __y1 = y1, __width = width - 1, __height = height - 1;
     CREATE_SHAPE(rect_t, rect);
